requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
boto3==1.34.0
django-storages==1.14.2
python-decouple==3.8
cryptography>=41.0.0
